+++ src/test/java/org/jabref/logic/citationstyle/CitationStyleTest.java
@@ -51,7 +51,7 @@ class CitationStyleTest {
     @ParameterizedTest
     @MethodSource
     void citationStylePresent(String cslFileName) {
-        Optional<CitationStyle> citationStyle = CSLStyleLoader.createCitationStyleFromFile(cslFileName);
+        Optional<CitationStyle> citationStyle = CSLStyleUtils.createCitationStyleFromFile(cslFileName);
         assertTrue(citationStyle.isPresent());
     }
 
@@ -68,8 +68,8 @@ class CitationStyleTest {
     @ParameterizedTest
     @MethodSource
     void titleMatches(String expectedTitle, String cslFileName) {
-        Optional<CitationStyle> citationStyle = CSLStyleLoader.createCitationStyleFromFile(cslFileName);
-        CSLStyleLoader.StyleInfo styleInfo = new CSLStyleLoader.StyleInfo(citationStyle.get().getTitle(), citationStyle.get().isNumericStyle());
+        Optional<CitationStyle> citationStyle = CSLStyleUtils.createCitationStyleFromFile(cslFileName);
+        CSLStyleUtils.StyleInfo styleInfo = new CSLStyleUtils.StyleInfo(citationStyle.get().getTitle(), citationStyle.get().isNumericStyle());
         assertEquals(expectedTitle, styleInfo.title());
     }
 
@@ -86,8 +86,8 @@ class CitationStyleTest {
     @ParameterizedTest
     @MethodSource
     void numericPropertyMatches(boolean expectedNumericNature, String cslFileName) {
-        Optional<CitationStyle> citationStyle = CSLStyleLoader.createCitationStyleFromFile(cslFileName);
-        CSLStyleLoader.StyleInfo styleInfo = new CSLStyleLoader.StyleInfo(citationStyle.get().getTitle(), citationStyle.get().isNumericStyle());
+        Optional<CitationStyle> citationStyle = CSLStyleUtils.createCitationStyleFromFile(cslFileName);
+        CSLStyleUtils.StyleInfo styleInfo = new CSLStyleUtils.StyleInfo(citationStyle.get().getTitle(), citationStyle.get().isNumericStyle());
         assertEquals(expectedNumericNature, styleInfo.isNumericStyle());
     }
